new songs, fix setlist too big

+++ software/src/actions/build_pdf.rs
@@ -1,10 +1,10 @@
 use crate::actions::build_lytex::build_lytex;
+use crate::actions::build_wav::build_wav;
 
+// use crate::generate::generate::generate_for_aws_lambda;
 use crate::helpers::helpers::song_of_booksong;
 use crate::helpers::io::{copy_file, read_to_string, read_to_vec_u8, write_string};
-use crate::model::model::{
-    Book, ELogType, LogItem, LogItemBook, LogItemSong, Song, StructureItemContent, World,
-};
+use crate::model::use_model as M;
 use regex::Regex;
 use sha2::{Digest, Sha256};
 use std::fs::File;
@@ -24,7 +24,7 @@ fn needs_rerun(path: PathBuf) -> Result<bool, Box<dyn std::error::Error>> {
 /// compute the digest of the sources of a song
 /// song.json ( for the structure )
 /// the lyrics, the additional tex files, the lilypond files
-fn compute_digest(song: &Song) -> Result<String, Box<dyn std::error::Error>> {
+fn compute_digest(song: &M::Song) -> Result<String, Box<dyn std::error::Error>> {
     let mut hasher = Sha256::new();
     hasher.update(serde_json::to_string(&song)?.as_bytes());
 
@@ -32,10 +32,10 @@ fn compute_digest(song: &Song) -> Result<String, Box<dyn std::error::Error>> {
         .structure
         .iter()
         .filter_map(|section| match &section.item {
-            StructureItemContent::ItemChords(chords) => Some(chords.section_id.clone()),
-            StructureItemContent::ItemHRule(_) => None,
-            StructureItemContent::ItemNewColumn => None,
-            StructureItemContent::ItemRef(xref) => Some(xref.section_id.clone()),
+            M::StructureItemContent::ItemChords(chords) => Some(chords.section_id.clone()),
+            M::StructureItemContent::ItemHRule(_) => None,
+            M::StructureItemContent::ItemNewColumn => None,
+            M::StructureItemContent::ItemRef(xref) => Some(xref.section_id.clone()),
         })
         .map(|id| {
             let mut p = PathBuf::from(&song.srcdir.clone());
@@ -67,7 +67,7 @@ fn compute_digest(song: &Song) -> Result<String, Box<dyn std::error::Error>> {
 
     let mut all_files: Vec<PathBuf> = vec![];
     all_files.append(&mut lyrics_files);
-    for f in vec!["song.json", "body.tex", "add.tikz"] {
+    for f in ["song.json", "body.tex", "add.tikz"] {
         all_files.push({
             let mut p = PathBuf::from(&song.srcdir.clone());
             p.push(f);
@@ -82,43 +82,46 @@ fn compute_digest(song: &Song) -> Result<String, Box<dyn std::error::Error>> {
         hasher.update(contents);
     }
     let result = hasher.finalize();
-    let ret = hex::encode(result.to_vec());
+    let ret = hex::encode(result);
     Ok(ret)
 }
 
 // computes the digest of the sources PLUS the output pdf file
 // Error if the pdf file does not exist
-pub fn compute_digest_ok(world: &World, song: &Song) -> Result<String, Box<dyn std::error::Error>> {
-    // log::info!("{}:{}", file!(), line!());
+pub fn compute_digest_ok(
+    world: &M::World,
+    song: &M::Song,
+) -> Result<String, Box<dyn std::error::Error>> {
+    // log::debug!("{}:{}", file!(), line!());
     let digest = compute_digest(song)?;
     let mut hasher = Sha256::new();
-    // log::info!("{}:{}", file!(), line!());
+    // log::debug!("{}:{}", file!(), line!());
     hasher.update(&digest);
-    // log::info!("{}:{}", file!(), line!());
+    // log::debug!("{}:{}", file!(), line!());
     {
         let mut p = PathBuf::from(&world.builddir);
         p.push("delivery");
-        // log::info!("{}:{}", file!(), line!());
+        // log::debug!("{}:{}", file!(), line!());
         p.push(format!("{}.pdf", &song.pdfname));
-        // log::info!("{}:{} {:?}", file!(), line!(), &p);
+        // log::debug!("{}:{} {:?}", file!(), line!(), &p);
         let contents = read_to_vec_u8(&p)?;
-        // log::info!("{}:{}", file!(), line!());
+        // log::debug!("{}:{}", file!(), line!());
         hasher.update(contents);
     }
-    // log::info!("{}:{}", file!(), line!());
+    // log::debug!("{}:{}", file!(), line!());
     let result = hasher.finalize();
-    // log::info!("{}:{}", file!(), line!());
-    let data = hex::encode(&result[..].to_vec());
-    // log::info!("{}:{} {}", file!(), line!(),&data);
+    // log::debug!("{}:{}", file!(), line!());
+    let data = hex::encode(&result[..]);
+    // log::debug!("{}:{} {}", file!(), line!(),&data);
     Ok(data)
 }
 
-fn pathbuf_ok_checksum(song: &Song) -> PathBuf {
+fn pathbuf_ok_checksum(song: &M::Song) -> PathBuf {
     let mut p = PathBuf::from(&song.builddir);
     p.push(".checksum_ok");
     p
 }
-// fn pathbuf_failed_checksum(song: &Song) -> PathBuf {
+// fn pathbuf_failed_checksum(song: & M::Song) -> PathBuf {
 //     let mut p = PathBuf::from(&song.builddir);
 //     p.push(".checksum_failed");
 //     p
@@ -137,10 +140,13 @@ fn __vec_are_equal<T: std::cmp::PartialEq>(v1: Vec<T>, v2: Vec<T>) -> bool {
 }
 
 /// true : at least one song needs rebuild
-fn needs_rebuild_book_ok(world: &World, book: &Book) -> Result<bool, Box<dyn std::error::Error>> {
+fn needs_rebuild_book_ok(
+    world: &M::World,
+    book: &M::Book,
+) -> Result<bool, Box<dyn std::error::Error>> {
     for bs in &book.songs {
         let song = song_of_booksong(world, bs)?;
-        if needs_rebuild_ok(&world, &song) {
+        if needs_rebuild_ok(world, &song) {
             return Ok(true);
         }
     }
@@ -148,18 +154,18 @@ fn needs_rebuild_book_ok(world: &World, book: &Book) -> Result<bool, Box<dyn std
 }
 
 /// true : song needs rebuild
-fn needs_rebuild_ok(world: &World, song: &Song) -> bool {
-    let digest = read_to_string(&pathbuf_ok_checksum(&song));
+fn needs_rebuild_ok(world: &M::World, song: &M::Song) -> bool {
+    let digest = read_to_string(&pathbuf_ok_checksum(song));
     match digest {
         Err(_) => true,
-        Ok(v1) => match compute_digest_ok(&world, &song) {
+        Ok(v1) => match compute_digest_ok(world, song) {
             Ok(v2) => v1 != v2,
             Err(_) => true,
         },
     }
 }
 
-// fn needs_rebuild_failed(song: &Song) -> Result<bool, Box<dyn std::error::Error>>  {
+// fn needs_rebuild_failed(song: & M::Song) -> Result<bool, Box<dyn std::error::Error>>  {
 //     let digest = read_to_string(&pathbuf_failed_checksum(&song));
 //     match digest {
 //         Err(_) => Ok(false),
@@ -171,78 +177,79 @@ fn needs_rebuild_ok(world: &World, song: &Song) -> bool {
 //     }
 // }
 
+/// build a pdf for a song, and also wav files if any
 pub async fn wrapped_build_pdf_song(
-    tx: Sender<LogItem>,
-    world: World,
-    song: Song,
+    tx: Sender<M::LogItem>,
+    world: M::World,
+    song: M::Song,
     force_rebuild: bool,
-) -> () {
+) {
     match build_pdf_song(tx, world.clone(), song.clone(), force_rebuild).await {
         Ok(()) => (),
         Err(e) => log::error!(
             "wrapped build pdf, {} {} ; {}",
             &song.author,
             &song.title,
-            e.to_string()
+            e
         ),
     }
 }
 
-pub async fn wrapped_build_pdf_book(tx: Sender<LogItem>, world: World, book: Book) -> () {
+pub async fn wrapped_build_pdf_book(tx: Sender<M::LogItem>, world: M::World, book: M::Book) {
     match build_pdf_book(tx, world.clone(), book.clone()).await {
         Ok(()) => (),
-        Err(e) => log::error!("wrapped build pdf, {} ; {}", &book.title, e.to_string()),
+        Err(e) => log::error!("wrapped build pdf, {} ; {}", &book.title, e),
     }
 }
 
 pub async fn build_pdf_song(
-    tx: Sender<LogItem>,
-    world: World,
-    song: Song,
+    tx: Sender<M::LogItem>,
+    world: M::World,
+    song: M::Song,
     force_rebuild: bool,
 ) -> Result<(), Box<dyn std::error::Error>> {
     log::info!("build pdf {} {}", &song.author, &song.title);
+
+    // generate_for_aws_lambda(&song.builddir)?;
+
     let p = pathbuf_ok_checksum(&song);
-    if force_rebuild {
-        if p.exists() {
-            std::fs::remove_file(p.as_path())?;
-        }
+    if force_rebuild && p.exists() {
+        std::fs::remove_file(p.as_path())?;
     }
     if !needs_rebuild_ok(&world, &song) {
-        log::info!("does not need rebuild {} {}", &song.author, &song.title);
-        let li = LogItem::Song(LogItemSong {
+        log::debug!("does not need rebuild {} {}", &song.author, &song.title);
+        let li = M::LogItem::Song(M::LogItemSong {
             author: song.author.clone(),
             title: song.title.clone(),
-            status: ELogType::NoNeedSuccess,
+            status: M::ELogType::NoNeedSuccess,
         });
         tx.send(li).await?;
         return Ok(());
     }
-
     // if needs_rebuild_failed(&song)? {
-    //     let li = LogItem::Song(LogItemSong {
+    //     let li = LogItem::Song(M::LogItemSong {
     //         author: song.author.clone(),
     //         title: song.title.clone(),
-    //         status: ELogType::NoNeedFailed,
+    //         status:  M::ELogType::NoNeedFailed,
     //     });
     //     tx.send(li).await.unwrap();
     //     return Ok(());
     // }
 
-    let li = LogItem::Song(LogItemSong {
+    let li = M::LogItem::Song(M::LogItemSong {
         author: song.author.clone(),
         title: song.title.clone(),
-        status: ELogType::Started,
+        status: M::ELogType::Started,
     });
     tx.send(li).await?;
 
     let mut success: bool = true;
 
     for lyfile in &song.lilypondfiles {
-        let li = LogItem::Song(LogItemSong {
+        let li = M::LogItem::Song(M::LogItemSong {
             author: song.author.clone(),
             title: song.title.clone(),
-            status: ELogType::Lilypond(lyfile.clone()),
+            status: M::ELogType::Lilypond(lyfile.clone()),
         });
         tx.send(li).await?;
         match build_lytex(song.clone(), lyfile.clone()).await {
@@ -251,28 +258,69 @@ pub async fn build_pdf_song(
         }
     }
 
+    for wavfile in &song.wavfiles {
+        let li = M::LogItem::Song(M::LogItemSong {
+            author: song.author.clone(),
+            title: song.title.clone(),
+            status: M::ELogType::Wav(wavfile.clone()),
+        });
+        tx.send(li).await?;
+        match build_wav(song.clone(), wavfile.clone()).await {
+            Ok(()) => (),
+            Err(e) => {
+                log::error!("{e:?}");
+                success = false
+            }
+        }
+    }
+
+    {
+        let _child = Command::new("mktexlsr")
+            //.env("HOME", song.builddir.to_str().unwrap())
+            .kill_on_drop(true)
+            // .stdout(Stdio::piped())
+            // .stdout(fout)
+            .current_dir(&song.builddir)
+            .spawn()?;
+    }
+
+    {
+        let _child = Command::new("mktexfmt")
+            .arg("user")
+            //.env("HOME", song.builddir.to_str().unwrap())
+            .kill_on_drop(true)
+            // .stdout(Stdio::piped())
+            // .stdout(fout)
+            .current_dir(&song.builddir)
+            .spawn()?;
+    }
+
     let mut count = 1;
     loop {
-        let fout = File::create(PathBuf::from("lualatex.log"))?;
+        let mut p: PathBuf = PathBuf::from(&song.builddir);
+        p.push("lualatex.log");
+        let fout = File::create(p)?;
         if !success {
             break;
         };
-        let li = LogItem::Song(LogItemSong {
+        let li = M::LogItem::Song(M::LogItemSong {
             author: song.author.clone(),
             title: song.title.clone(),
-            status: ELogType::Lualatex(count as u32),
+            status: M::ELogType::Lualatex(count as u32),
         });
         tx.send(li).await?;
+        log::info!("{}:{} running lualatex", file!(), line!());
         let child = Command::new("lualatex")
             .arg("--interaction=nonstopmode")
             .arg("main.tex")
+            //.env("HOME", song.builddir.to_str().unwrap())
             .kill_on_drop(true)
             // .stdout(Stdio::piped())
             .stdout(fout)
             .current_dir(&song.builddir)
             .spawn()?;
 
-        log::info!(
+        log::debug!(
             "{}:{} run lualatex for {} {}",
             file!(),
             line!(),
@@ -280,17 +328,22 @@ pub async fn build_pdf_song(
             &song.title
         );
         let output = &child.wait_with_output().await?;
-
         if !(output.status.success()) {
             success = false;
             log::error!("lualatex failed for {} {}", &song.author, &song.title);
+            // return Err(Box::new(
+            //     std::io::Error::new(
+            //         std::io::ErrorKind::Other,
+            //         format!("lualatex failed for {} {}", &song.author, &song.title),
+            //     ),
+            // ));
             break;
         }
         if !needs_rerun(song.builddir.clone())? {
             success = true;
             break;
         }
-        count = count + 1;
+        count += 1;
     }
 
     {
@@ -306,29 +359,30 @@ pub async fn build_pdf_song(
         // copy_file(&pfrom, &pto)?;
 
         let fout = File::create(PathBuf::from("lualatex.log"))?;
-        let li = LogItem::Song(LogItemSong {
+        let li = M::LogItem::Song(M::LogItemSong {
             author: song.author.clone(),
             title: song.title.clone(),
-            status: ELogType::Ps2pdf,
+            status: M::ELogType::Ps2pdf,
         });
         tx.send(li).await?;
         let child = Command::new("ps2pdf")
             .arg("main.pdf")
             .arg(pto.to_str().unwrap())
+            //.env("HOME", song.builddir.to_str().unwrap())
             .kill_on_drop(true)
             // .stdout(Stdio::piped())
             .stdout(fout)
             .current_dir(&song.builddir)
             .spawn()?;
 
-        log::info!(
+        log::debug!(
             "{}:{} run ps2pdf for {} {}",
             file!(),
             line!(),
             &song.author,
             &song.title
         );
-        // log::info!("{}:{} {:?}", file!(), line!(), &child);
+        // log::debug!("{}:{} {:?}", file!(), line!(), &child);
         let output = &child.wait_with_output().await?;
         if !(output.status.success()) {
             success = false;
@@ -337,12 +391,12 @@ pub async fn build_pdf_song(
     }
 
     if !success {
-        let li = LogItem::Song(LogItemSong {
+        let li = M::LogItem::Song(M::LogItemSong {
             author: song.author.clone(),
             title: song.title.clone(),
-            status: ELogType::Failed,
+            status: M::ELogType::Failed,
         });
-        log::info!(
+        log::debug!(
             "tx.send final error result for {} {}",
             &song.author,
             &song.title
@@ -354,17 +408,17 @@ pub async fn build_pdf_song(
     if success {
         // write ok checksum
         let checksum = compute_digest_ok(&world, &song)?;
-        let _x = write_string(&pathbuf_ok_checksum(&song), &checksum)?;
+        write_string(&pathbuf_ok_checksum(&song), &checksum)?;
         // assert!(needs_rebuild_ok(&song) == false);
     }
 
-    let li = LogItem::Song(LogItemSong {
+    let li = M::LogItem::Song(M::LogItemSong {
         author: song.author.clone(),
         title: song.title.clone(),
         status: if success {
-            ELogType::Success
+            M::ELogType::Success
         } else {
-            ELogType::Failed
+            M::ELogType::Failed
         },
     });
     log::info!(
@@ -377,24 +431,24 @@ pub async fn build_pdf_song(
 }
 
 async fn build_pdf_book(
-    tx: Sender<LogItem>,
-    world: World,
-    book: Book,
+    tx: Sender<M::LogItem>,
+    world: M::World,
+    book: M::Book,
 ) -> Result<(), Box<dyn std::error::Error>> {
-    log::info!("build book {}", &book.title);
+    log::info!("{}:{} build book {}", file!(), line!(), &book.title);
     loop {
         if !needs_rebuild_book_ok(&world, &book)? {
             break;
         }
-        log::info!(
+        log::debug!(
             "wait for building book {} ... some songs are not done yet",
             book.title
         );
         let _x = tokio::time::sleep(Duration::from_secs(5)).await;
     }
-    let li = LogItem::Book(LogItemBook {
+    let li = M::LogItem::Book(M::LogItemBook {
         title: book.title.clone(),
-        status: ELogType::Started,
+        status: M::ELogType::Started,
     });
     tx.send(li).await?;
 
@@ -406,9 +460,9 @@ async fn build_pdf_book(
             break;
         };
         let fout = File::create(PathBuf::from("lualatex.log"))?;
-        let li = LogItem::Book(LogItemBook {
+        let li = M::LogItem::Book(M::LogItemBook {
             title: book.title.clone(),
-            status: ELogType::Lualatex(count as u32),
+            status: M::ELogType::Lualatex(count as u32),
         });
         tx.send(li).await?;
         let child = Command::new("lualatex")
@@ -420,7 +474,7 @@ async fn build_pdf_book(
             .current_dir(&book.builddir)
             .spawn()?;
 
-        log::info!(
+        log::debug!(
             "{}:{} run lualatex for book {}",
             file!(),
             line!(),
@@ -436,10 +490,10 @@ async fn build_pdf_book(
             success = true;
             break;
         }
-        count = count + 1;
+        count += 1;
     }
 
-    log::info!(
+    log::debug!(
         "{}:{} success : {} ; book {}",
         file!(),
         line!(),
@@ -448,11 +502,11 @@ async fn build_pdf_book(
     );
 
     if !success {
-        let li = LogItem::Book(LogItemBook {
+        let li = M::LogItem::Book(M::LogItemBook {
             title: book.title.clone(),
-            status: ELogType::Failed,
+            status: M::ELogType::Failed,
         });
-        log::info!("tx.send final error result for {} ", &book.title);
+        log::debug!("tx.send final error result for {} ", &book.title);
         tx.send(li).await?;
         return Ok(());
     }
@@ -463,9 +517,9 @@ async fn build_pdf_book(
         pto.push(format!("{}.pdf", &book.pdfname));
 
         let fout = File::create(PathBuf::from("lualatex.log"))?;
-        let li = LogItem::Book(LogItemBook {
+        let li = M::LogItem::Book(M::LogItemBook {
             title: book.title.clone(),
-            status: ELogType::Ps2pdf,
+            status: M::ELogType::Ps2pdf,
         });
         tx.send(li).await?;
         let child = Command::new("ps2pdf")
@@ -477,8 +531,8 @@ async fn build_pdf_book(
             .current_dir(&book.builddir)
             .spawn()?;
 
-        log::info!("{}:{} run ps2pdf for {} ", file!(), line!(), &book.title);
-        log::info!("{}:{} {:?}", file!(), line!(), &child);
+        log::debug!("{}:{} run ps2pdf for {} ", file!(), line!(), &book.title);
+        log::debug!("{}:{} {:?}", file!(), line!(), &child);
         let output = &child.wait_with_output().await?;
         if !(output.status.success()) {
             success = false;
@@ -486,15 +540,15 @@ async fn build_pdf_book(
         }
     }
 
-    let li = LogItem::Book(LogItemBook {
+    let li = M::LogItem::Book(M::LogItemBook {
         title: book.title.clone(),
         status: if success {
-            ELogType::Success
+            M::ELogType::Success
         } else {
-            ELogType::Failed
+            M::ELogType::Failed
         },
     });
-    log::info!("tx.send final result for book {}", &book.title);
+    log::debug!("tx.send final result for book {}", &book.title);
     tx.send(li).await?;
 
     Ok(())
